separate a module to set renderer for featurecollection
