+++ src/FeatureCollection/Renderer.js
@@ -1,0 +1,46 @@
+import { classBreaksRenderer } from 'esri-leaflet-renderers/src/Renderers/ClassBreaksRenderer';
+import { uniqueValueRenderer } from 'esri-leaflet-renderers/src/Renderers/UniqueValueRenderer';
+import { simpleRenderer } from 'esri-leaflet-renderers/src/Renderers/SimpleRenderer';
+
+export function setRenderer (layerDefinition, layer) {
+  var rend;
+  var rendererInfo = layer.renderer;
+
+  var options = {};
+
+  if (layer.options.pane) {
+    options.pane = layer.options.pane;
+  }
+  if (layerDefinition.drawingInfo.transparency) {
+    options.layerTransparency = layerDefinition.drawingInfo.transparency;
+  }
+  if (layer.options.style) {
+    options.userDefinedStyle = layer.options.style;
+  }
+
+  switch (rendererInfo.type) {
+    case 'classBreaks':
+      layer._checkForProportionalSymbols(layerDefinition.geometryType, rendererInfo);
+      if (layer._hasProportionalSymbols) {
+        layer._createPointLayer();
+        var pRend = classBreaksRenderer(rendererInfo, options);
+        pRend.attachStylesToLayer(layer._pointLayer);
+        options.proportionalPolygon = true;
+      }
+      rend = classBreaksRenderer(rendererInfo, options);
+      break;
+    case 'uniqueValue':
+      console.log(rendererInfo, options);
+      rend = uniqueValueRenderer(rendererInfo, options);
+      break;
+    default:
+      rend = simpleRenderer(rendererInfo, options);
+  }
+  rend.attachStylesToLayer(layer);
+}
+
+export var Renderer = {
+  setRenderer: setRenderer
+};
+
+export default Renderer;
